clarify stability a bit
[skip ci]

Clarify the intent to help individual users working on an open source project using `elastic` and trying to keep up to date with the changes. I think this is better use of time right now than worrying too much about changelogs. Within a few releases we should be able to lean on the Changelog to communicate changes and describe an upgrade path more.